Extract a page type out of the andrew server itself.
The server clearly had some concepts inside it that were about page
content; I just tried to discover a boundary and help it to be clearer.

The Page conceptually deals with a file and its content; the Server
deals with the file system and getting content streamed to the end user.