Adds polling for whether the server is ready.
I can make these tests work by choosing the :0 syntax, so I thought the
error was caused by dns resolution. After some tinkering, though, it
looks like the error may be that the server starts and stops with
different timing on macos vs linux; entering a synchronisation loop
should fix it if that is hte issue.