Fixes the test that verifies andrew-publish-time parsing.
This was in the Page constructor, which performs that parsing for a new
Page. In that function, I do not pass a timezone in to the date time
parser, as it seems irrelevant.

THe tests, though, had their test data created with a time zone in
there. THat meant I was comparing dates across timezones
unintentionally.