Extracts logic from constructor into helper function.
This follows the same pattern as getTitle etc. I do anticipate
needing to extract more information in future from meta tags, but
for now this functionality is only used to get the publish date.

This was to support an important goal: since I put the template tag
rendering into Andrew, it has been restricted to index.html pages.

Removing that restriction revealed a fairly unpleasant truth: the
constructor NewPage was being called recursively within itself, as hidden
inside the GetSiblingsAndChildren function is a call to the NewPage
constructor.

This had worked because GetSiblingsAndCHildren was only called for
index.html files, but itself refused to process index.html because
I never wanted a link generated that was to an index.html page.

I removed the restriction for the constructor to only call
GetSiblingsAndChildren for index.html pages, and refactored that
function to no longer call NewPage(). Instead it creates an instance
of a page using the struct.

Now there are de facto two separate implementations of meeting the
preconditions for creating a new page, but that can be cleaned up when
it comes up.